{{cookiecutter.project_slug}}/pyproject.toml
[tool.poetry]
name = "{{cookiecutter.project_slug}}"
version = "0.1.0"
description = ""
authors = ["{{cookiecutter.author}} <{{cookiecutter.email}}>"]

[tool.poetry.dependencies]
python = "^3.7"
pydantic = {version = "^1.7.3"}
attrs = "^20.3.0"
sqlalchemy = "^1.3.22"
fastapi = "^0.63.0"
psycopg2-binary = "^2.8.6"
fastapi-users = {extras = ["sqlalchemy"], version = "^5.1.0"}
databases = {extras = ["postgresql"], version = "^0.4.1"}

[tool.poetry.dev-dependencies]
pytest = "^6.2.1"
requests = "^2.25.1"
pytest-dotenv = "^0.5.2"
factory-boy = "^3.2.0"
asyncio = "^3.4.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.pytest.ini_options]
markers = [
    "end2end",
]

env_files = [
    ".dev.env",
    ".env",
]
